+++ Runtime/Attributes/ComponentGroupAttribute.cs
@@ -1,41 +1,6 @@
-﻿#if ENABLE_IL2CPP
-#define INLINE_METHODS
-#endif
-using System.Reflection;
+﻿using System.Reflection;
 
 namespace ME.ECS {
-
-    using ME.ECS.Collections;
-
-    public class IsBitmask : System.Attribute { }
-
-    public enum ComponentLifetime : byte {
-
-        Infinite = 0,
-
-        NotifyAllSystemsBelow = 1,
-        NotifyAllSystems = 2,
-
-    }
-
-    public enum StorageType : byte {
-
-        Default = 0,
-        NoState = 1,
-
-    }
-
-    public class ComponentOrderAttribute : System.Attribute {
-
-        public int order;
-
-        public ComponentOrderAttribute(int order) {
-
-            this.order = order;
-
-        }
-
-    }
 
     public enum GroupColor {
 
@@ -118,59 +83,6 @@ namespace ME.ECS {
             }
 
         }
-
-    }
-
-    public interface IComponentBase { }
-
-    public interface IStructComponent : IComponentBase { }
-
-    public interface IComponent : IStructComponent { }
-
-    public interface IComponentRuntime { }
-    
-    public interface IComponentOneShot : IComponentBase, IComponentRuntime { }
-
-    public interface IVersioned : IComponentBase { }
-
-    public interface IComponentBlittable : IComponentBase {}
-    
-    #if !COMPONENTS_VERSION_NO_STATE_DISABLED
-    public interface IVersionedNoState : IComponentBase { }
-    #endif
-
-    public interface IComponentDisposable : IComponentBase {
-
-        void OnDispose();
-
-    }
-
-    public interface ICopyableBase { }
-
-    public interface IStructCopyable<T> : IComponent, ICopyableBase where T : IStructCopyable<T> {
-
-        void CopyFrom(in T other);
-        void OnRecycle();
-
-    }
-
-    public interface ICopyable<T> : IStructCopyable<T> where T : IStructCopyable<T> {
-
-    }
-
-    #if !SHARED_COMPONENTS_DISABLED
-    public interface IComponentShared : IComponentBase { }
-
-    public interface ISharedGroups {
-
-        System.Collections.Generic.ICollection<uint> GetGroups();
-
-    }
-    #endif
-
-    public interface IStructComponentsContainer {
-
-        BufferArray<StructRegistryBase> GetAllRegistries();
 
     }