+++ Runtime/Interfaces/IComponentShared.cs
@@ -1,0 +1,13 @@
+namespace ME.ECS {
+
+    #if !SHARED_COMPONENTS_DISABLED
+    public interface IComponentShared : IComponentBase { }
+
+    public interface ISharedGroups {
+
+        System.Collections.Generic.ICollection<uint> GetGroups();
+
+    }
+    #endif
+
+}